+++ evealert/__init__.py
@@ -1,4 +1,4 @@
 """Initialize the app"""
 
-__version__ = "6.3.26"
+__version__ = "6.3.27"
 __title__ = "EVE Alert"